Carbon sequestration via afforestation
The project is devoted to activities on protection and management of the afforested degraded agricultural land in Siberian settlements in Altai Kray, Russian Federation. The project covers 9489,37 hectares of land, that have the status of the non-forested area. Project activities are aimed at sequestration of CO2 via the creation of new forest (carbon-absorbing forest planting) on the post-agrogenic crude soils of Ob-Chumysh rivers through stopping active agricultural activities and organization of forest protection and management measures. Also algorithms for estimation of the carbon absorption in forest ecosystems on the local level, and through that, implementation of the JI project activities corresponding to atr. 6 of Kyoto Protocol is developed; active management and protection of the forest areas not included in the State Forest Fund are engaged.